+++ CHANGELOG.md
@@ -6,5 +6,6 @@ v0.2.0
 
 * Added editable attribute to TextArea (Allows selection, but not alteration).
 * Added Element#font= and :font option.
+* Added Gosu::Color#colorize to use when using in-line text styling.
 * Managed layout of entities and XML tags (Used by Gosu) in TextArea text better (tags still don't like newlines inside them).
 * Changed license from LGPL to MIT.